+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 *   Fixed multimeter value to work on units and names #89
 *   Fixed display artifact on x-axis timescale #97
 *   Added elapse time display formatting option as Units → elapsed_time #88
+*   Added click on multimeter accumulate text to copy.
 
 
 ## 0.9.2